Connect Board to redux store in order to fetch and display data for current selected board
